--- src/actions/acitonCreator.tsx
@@ -1,9 +1,0 @@
-import { sampleAction } from './SampleAction';
-
-class AppActionCreator {
-  public sampleAction = sampleAction;
-}
-
-const appActionCreator = new AppActionCreator();
-
-export default appActionCreator;
