--- src/actions/actions.tsx
@@ -1,5 +1,0 @@
-enum Actions {
-  SAMPLE_ACTION = 'SAMPLE_ACTION'
-}
-
-export default Actions;
