+++ src/reducer/rootReducer.tsx
@@ -1,0 +1,8 @@
+import { combineReducers } from 'redux';
+import SampleReducer from './SampleReducer';
+
+const rootReducer = combineReducers({
+  SampleReducer,
+});
+
+export default rootReducer;
